% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pyphlawd}
\alias{pyphlawd}
\alias{setup_clade}
\title{pyphlawd}
\usage{
setup_clade(taxon, db, outdir, logfile = "log.txt", configfile = NULL)
}
\arguments{
\item{taxon}{Taxon}

\item{db}{db}

\item{outdir}{outdir}

\item{logfile}{logfile}

\item{configfile}{configfile}
}
\description{
Run pyphlawd
}
\examples{
library(outsider)


}
